\name{ri_iv_gridsearch}
\alias{ri_iv_gridsearch}
\docType{package}
\title{
  Grid search for CIs using RI.
}
\description{
  Simple function using Rcpp
}
\usage{
  ri_iv_gridsearch(y,x,z,
  grid,
  allowInf=FALSE,
  level=0.9,
  plot=FALSE,
  refinementsteps=1,
  refinementsresolution=11,
  seed=1234,
  gridres=NULL,
  r=1000,
  verbose=FALSE,
  wilcox=FALSE
}
\examples{
}
