\name{riiv-package}
\alias{riiv-package}
\alias{riiv}
\docType{package}
\title{
  Code for Randomization Inference and Instrumental Variables
}
\description{
  Some functions to conduct RI with IV and to run simulations related to this
}
\details{
  No more details at this point
}
\author{
  SH and MW
}
\references{
}
\keyword{ package, randomization inference, instrumental variables }
\seealso{

}
\examples{

}
